refactor: 상품 도메인 객체 금액 원시값 포장, 데이터타입 변경

+++ src/main/java/cart/dto/ProductRequest.java
@@ -2,13 +2,13 @@ package cart.dto;
 
 public class ProductRequest {
     private String name;
-    private int price;
+    private long price;
     private String imageUrl;
 
     public ProductRequest() {
     }
 
-    public ProductRequest(String name, int price, String imageUrl) {
+    public ProductRequest(String name, long price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -18,7 +18,7 @@ public class ProductRequest {
         return name;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
